SWT-320 Rename few files back to *sweety*
