+++ src/index.ts
@@ -1,9 +1,9 @@
 export type { Compare } from "./utils"
-export { Impulse } from "./Impulse"
+export { Impulse } from "./Sweety"
 export { batch } from "./batch"
 export { watch } from "./watch"
-export { useImpulse } from "./useImpulse"
-export { useImpulseValue } from "./useImpulseValue"
+export { useImpulse } from "./useSweety"
+export { useImpulseValue } from "./useSweetyState"
 export { useImpulseMemo } from "./useImpulseMemo"
 export { useImpulseEffect, useImpulseLayoutEffect } from "./useImpulseEffects"
 export { useWatchImpulse } from "./useWatchImpulse"
