+++ tests/hooks-in-components/watching-multiple-stores.spec.tsx
