+++ tests/useWatchSweety/illegal-usage.spec.tsx
